backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
httpx>=0.24.0
pydantic>=2.0.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
sendgrid>=6.11.0
websockets>=12.0
groq>=0.4.0
